+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog of lizard-elevationprofile
 ------------------
 
 - Show elevation profile in Box awesome for lizard day.
-- Fixed iPad bug when drawing elevation profiles
+- Fixed iPad bug when drawing elevation profiles.
+- Fix js for IE.
 
 
 0.2.2 (2013-06-04)